Separate compilation units of sparse tensor converters
